# python-server/openapi_tmp.json
{"openapi":"3.1.0","info":{"title":"Photo Organizer API","version":"0.1.0"},"paths":{"/process-image/":{"post":{"summary":"Detect Tags","description":"Upload an image and return all detected object tags (YOLO classes above threshold).","operationId":"detect_tags_process_image__post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_detect_tags_process_image__post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/folders/":{"get":{"summary":"List Folders","description":"Return a list of subfolder names under TARGET_FOLDER.","operationId":"list_folders_folders__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/folders/{folder_name}/":{"get":{"summary":"List Folder Files","description":"Return a list of file URLs for the requested subfolder.","operationId":"list_folder_files_folders__folder_name___get","parameters":[{"name":"folder_name","in":"path","required":true,"schema":{"type":"string","title":"Folder Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/all-organized-images/":{"get":{"summary":"List All Organized Images","description":"Return a list of all organized image URLs recursively.","operationId":"list_all_organized_images_all_organized_images__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/tags/{filename}/":{"get":{"summary":"Get Tags For File","description":"Return tags for a specific filename if present in DB.","operationId":"get_tags_for_file_tags__filename___get","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/all-organized-images-with-tags/":{"get":{"summary":"List All Organized Images With Tags","description":"Return a list of images and tags (uses server-side tag DB).","operationId":"list_all_organized_images_with_tags_all_organized_images_with_tags__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_detect_tags_process_image__post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_detect_tags_process_image__post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}